refactor: moved Organization module tests inside the module directory and updated configuration files
