+++ composer.json
@@ -36,7 +36,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Tests\\": "tests/"
+            "Tests\\": "tests/",
+            "Modules\\Organization\\Tests\\": "modules/Organization/Tests/"
         }
     },
     "scripts": {
